--- frontend/src/widgets/inputs/DateTimeInputWidget.tsx
@@ -1,627 +1,0 @@
-import * as React from 'react';
-import { useState, useCallback, useMemo } from 'react';
-import { Button } from '@/components/ui/button';
-import { Calendar } from '@/components/ui/calendar';
-import { Input } from '@/components/ui/input';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
-import { format } from 'date-fns';
-import { Calendar as CalendarIcon, Clock, X } from 'lucide-react';
-import { cn } from '@/lib/utils';
-import { useEventHandler } from '@/components/event-handler';
-import { inputStyles } from '@/lib/styles';
-import { InvalidIcon } from '@/components/InvalidIcon';
-import { Scales } from '@/types/scale';
-import {
-  dateTimeInputVariants,
-  dateTimeInputIconVariants,
-  dateTimeInputTextVariants,
-} from '@/components/ui/input/date-time-input-variants';
-
-type VariantType = 'Date' | 'DateTime' | 'Time';
-
-interface DateTimeInputWidgetProps {
-  id: string;
-  value?: string;
-  placeholder?: string;
-  disabled: boolean;
-  variant?: VariantType;
-  nullable?: boolean;
-  invalid?: string;
-  format?: string;
-  scale?: Scales;
-  'data-testid'?: string;
-}
-
-interface BaseVariantProps {
-  id: string;
-  value?: string;
-  placeholder?: string;
-  disabled: boolean;
-  nullable?: boolean;
-  invalid?: string;
-  format?: string;
-  scale?: Scales;
-  'data-testid'?: string;
-}
-
-interface DateVariantProps extends BaseVariantProps {
-  onDateChange: (date: Date | undefined) => void;
-}
-
-interface DateTimeVariantProps extends BaseVariantProps {
-  onDateChange: (date: Date | undefined) => void;
-  onTimeChange: (time: string) => void;
-}
-
-interface TimeVariantProps extends BaseVariantProps {
-  onTimeChange: (time: string) => void;
-}
-
-const DateVariant: React.FC<DateVariantProps> = ({
-  value,
-  placeholder,
-  disabled,
-  nullable,
-  invalid,
-  onDateChange,
-  format: formatProp,
-  scale = Scales.Medium,
-  'data-testid': dataTestId,
-}) => {
-  const [open, setOpen] = useState(false);
-  const date = value ? new Date(value) : undefined;
-  const showClear = nullable && !disabled && value != null && value !== '';
-
-  const handleClear = (e?: React.MouseEvent) => {
-    e?.preventDefault();
-    e?.stopPropagation();
-    onDateChange(undefined);
-  };
-
-  const handleSelect = useCallback(
-    (selectedDate: Date | undefined) => {
-      onDateChange(selectedDate);
-      setOpen(false);
-    },
-    [onDateChange]
-  );
-
-  return (
-    <div className="relative w-full select-none">
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>
-          <Button
-            disabled={disabled}
-            variant="outline"
-            className={cn(
-              dateTimeInputVariants({ scale }),
-              !date && 'text-muted-foreground',
-              invalid && inputStyles.invalidInput,
-              disabled && 'cursor-not-allowed',
-              showClear && invalid
-                ? 'pr-16'
-                : showClear || invalid
-                  ? 'pr-8'
-                  : ''
-            )}
-            data-testid={dataTestId}
-          >
-            <CalendarIcon
-              className={cn(
-                'mr-2 shrink-0',
-                dateTimeInputIconVariants({ scale })
-              )}
-            />
-            <span
-              className={cn(
-                'truncate',
-                dateTimeInputTextVariants({ scale }),
-                !date && 'text-muted-foreground'
-              )}
-            >
-              {date
-                ? format(date, formatProp || 'yyyy-MM-dd')
-                : placeholder || 'Pick a date'}
-            </span>
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent className="w-auto p-0" align="start">
-          <Calendar
-            mode="single"
-            selected={date}
-            onSelect={handleSelect}
-            initialFocus
-            scale={scale}
-          />
-        </PopoverContent>
-      </Popover>
-      {/* Icons absolutely positioned */}
-      {(showClear || invalid) && (
-        <div className="absolute right-2 top-1/2 -translate-y-1/2 flex flex-row items-center gap-1">
-          {showClear && (
-            <button
-              type="button"
-              tabIndex={-1}
-              aria-label="Clear"
-              onClick={handleClear}
-              className="p-1 rounded hover:bg-accent focus:outline-none cursor-pointer"
-            >
-              <X
-                className={cn(
-                  dateTimeInputIconVariants({ scale }),
-                  'text-muted-foreground hover:text-foreground'
-                )}
-              />
-            </button>
-          )}
-          {/* Invalid icon - rightmost */}
-          {invalid && (
-            <InvalidIcon message={invalid} className="pointer-events-auto" />
-          )}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const DateTimeVariant: React.FC<DateTimeVariantProps> = ({
-  value,
-  placeholder,
-  disabled,
-  nullable,
-  invalid,
-  onDateChange,
-  onTimeChange,
-  format: formatProp,
-  scale = Scales.Medium,
-  'data-testid': dataTestId,
-}) => {
-  const [open, setOpen] = useState(false);
-  const date = useMemo(() => (value ? new Date(value) : undefined), [value]);
-  const showClear = nullable && !disabled && value != null && value !== '';
-
-  const handleClear = (e?: React.MouseEvent) => {
-    e?.preventDefault();
-    e?.stopPropagation();
-    onDateChange(undefined);
-  };
-
-  // Use local state for the time input to make it uncontrolled
-  const [localTimeValue, setLocalTimeValue] = useState(() => {
-    if (date) {
-      return format(date, formatProp || 'HH:mm:ss');
-    }
-    return '00:00:00';
-  });
-
-  // Track if user is actively editing the time input
-  const [isEditingTime, setIsEditingTime] = useState(false);
-
-  // Update local time when date changes, but only if user is not actively editing
-  React.useEffect(() => {
-    if (!isEditingTime) {
-      if (date) {
-        const newTimeValue = format(date, formatProp || 'HH:mm:ss');
-        setLocalTimeValue(newTimeValue);
-      } else if (nullable) {
-        // When nullable and no date, keep input empty instead of defaulting to '00:00:00'
-        setLocalTimeValue('');
-      } else {
-        setLocalTimeValue('00:00:00');
-      }
-    }
-  }, [date, formatProp, isEditingTime, nullable]);
-
-  const handleDateSelect = useCallback(
-    (selectedDate: Date | undefined) => {
-      if (selectedDate) {
-        // Preserve the time when selecting a new date
-        const currentTime = date ? date : new Date();
-        const newDateTime = new Date(selectedDate);
-        newDateTime.setHours(
-          currentTime.getHours(),
-          currentTime.getMinutes(),
-          currentTime.getSeconds()
-        );
-        onDateChange(newDateTime);
-      } else {
-        onDateChange(undefined);
-      }
-      // Do not close the popover here, allow user to pick time
-    },
-    [date, onDateChange]
-  );
-
-  const handleTimeChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const newTimeValue = e.target.value;
-      setLocalTimeValue(newTimeValue);
-      setIsEditingTime(true);
-    },
-    []
-  );
-
-  const handleTimeBlur = useCallback(() => {
-    setIsEditingTime(false);
-    // When time input loses focus, update the parent
-    if (date && localTimeValue) {
-      const [hours, minutes, seconds] = localTimeValue.split(':').map(Number);
-      const newDateTime = new Date(date);
-      newDateTime.setHours(hours, minutes, seconds);
-      onDateChange(newDateTime);
-    } else if (localTimeValue) {
-      // If no date is selected, create a new date with the selected time
-      const [hours, minutes, seconds] = localTimeValue.split(':').map(Number);
-      const newDateTime = new Date();
-      newDateTime.setHours(hours, minutes, seconds);
-      onDateChange(newDateTime);
-    }
-    onTimeChange(localTimeValue);
-  }, [date, localTimeValue, onDateChange, onTimeChange]);
-
-  const handleTimeKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      // When user presses Enter, update the parent
-      if (e.key === 'Enter') {
-        setIsEditingTime(false);
-        if (date && localTimeValue) {
-          const [hours, minutes, seconds] = localTimeValue
-            .split(':')
-            .map(Number);
-          const newDateTime = new Date(date);
-          newDateTime.setHours(hours, minutes, seconds);
-          onDateChange(newDateTime);
-        } else if (localTimeValue) {
-          // If no date is selected, create a new date with the selected time
-          const [hours, minutes, seconds] = localTimeValue
-            .split(':')
-            .map(Number);
-          const newDateTime = new Date();
-          newDateTime.setHours(hours, minutes, seconds);
-          onDateChange(newDateTime);
-        }
-        onTimeChange(localTimeValue);
-      }
-    },
-    [date, localTimeValue, onDateChange, onTimeChange]
-  );
-
-  const handleTimeFocus = useCallback(() => {
-    setIsEditingTime(true);
-  }, []);
-
-  return (
-    <div className="relative w-full select-none">
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>
-          <Button
-            disabled={disabled}
-            variant="outline"
-            className={cn(
-              dateTimeInputVariants({ scale }),
-              !date && 'text-muted-foreground',
-              invalid && inputStyles.invalidInput,
-              disabled && 'cursor-not-allowed',
-              showClear && invalid
-                ? 'pr-16'
-                : showClear || invalid
-                  ? 'pr-8'
-                  : ''
-            )}
-            data-testid={dataTestId}
-          >
-            <CalendarIcon
-              className={cn(
-                'mr-2 shrink-0',
-                dateTimeInputIconVariants({ scale })
-              )}
-            />
-            <Clock
-              className={cn(
-                'mr-2 shrink-0',
-                dateTimeInputIconVariants({ scale })
-              )}
-            />
-            <span
-              className={cn(
-                'truncate',
-                dateTimeInputTextVariants({ scale }),
-                !date && 'text-muted-foreground'
-              )}
-            >
-              {date
-                ? format(date, formatProp || 'yyyy-MM-dd')
-                : placeholder || 'Pick a date & time'}
-            </span>
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent className="w-auto p-0" align="start">
-          <div className="flex flex-col gap-2 p-2">
-            <Calendar
-              mode="single"
-              selected={date}
-              onSelect={handleDateSelect}
-              initialFocus
-              scale={scale}
-            />
-            <div className="flex items-center gap-2">
-              <Clock
-                className={cn(
-                  dateTimeInputIconVariants({ scale }),
-                  'text-muted-foreground'
-                )}
-              />
-              <Input
-                type="time"
-                step="1"
-                value={localTimeValue}
-                onChange={handleTimeChange}
-                onFocus={handleTimeFocus}
-                onBlur={handleTimeBlur}
-                onKeyDown={handleTimeKeyDown}
-                disabled={disabled}
-                className={cn(
-                  'bg-transparent appearance-none [&::-webkit-calendar-picker-indicator]:hidden',
-                  dateTimeInputTextVariants({ scale }),
-                  invalid && inputStyles.invalidInput
-                )}
-                data-testid={dataTestId ? `${dataTestId}-time` : undefined}
-              />
-            </div>
-          </div>
-        </PopoverContent>
-      </Popover>
-      {/* Icons absolutely positioned */}
-      {(showClear || invalid) && (
-        <div className="absolute right-2 top-1/2 -translate-y-1/2 flex flex-row items-center gap-1">
-          {showClear && (
-            <button
-              type="button"
-              tabIndex={-1}
-              aria-label="Clear"
-              onClick={handleClear}
-              className="p-1 rounded hover:bg-accent focus:outline-none cursor-pointer"
-            >
-              <X
-                className={cn(
-                  dateTimeInputIconVariants({ scale }),
-                  'text-muted-foreground hover:text-foreground'
-                )}
-              />
-            </button>
-          )}
-          {/* Invalid icon - rightmost */}
-          {invalid && (
-            <InvalidIcon message={invalid} className="pointer-events-auto" />
-          )}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const TimeVariant: React.FC<TimeVariantProps> = ({
-  value,
-  placeholder,
-  disabled,
-  nullable,
-  invalid,
-  onTimeChange,
-  scale = Scales.Medium,
-  'data-testid': dataTestId,
-}) => {
-  // Use local state for the input value to make it uncontrolled
-  const [localTimeValue, setLocalTimeValue] = useState(() => {
-    if (value) {
-      // Parse the value to get time in HH:mm:ss format
-      try {
-        const date = new Date(value);
-        if (!isNaN(date.getTime())) {
-          return format(date, 'HH:mm:ss');
-        }
-      } catch {
-        // If parsing fails, try to use the value directly if it looks like a time
-        if (
-          typeof value === 'string' &&
-          /^\d{1,2}:\d{2}(:\d{2})?$/.test(value)
-        ) {
-          return value.length <= 5 ? value + ':00' : value;
-        }
-      }
-    }
-    // When nullable and no value, return empty string to show placeholder
-    return nullable ? '' : '00:00:00';
-  });
-
-  // Update local state when value prop changes (from parent)
-  React.useEffect(() => {
-    if (value) {
-      try {
-        const date = new Date(value);
-        if (!isNaN(date.getTime())) {
-          const newTimeValue = format(date, 'HH:mm:ss');
-          setLocalTimeValue(newTimeValue);
-        }
-      } catch {
-        // If parsing fails, try to use the value directly if it looks like a time
-        if (
-          typeof value === 'string' &&
-          /^\d{1,2}:\d{2}(:\d{2})?$/.test(value)
-        ) {
-          const newTimeValue = value.length <= 5 ? value + ':00' : value;
-          setLocalTimeValue(newTimeValue);
-        }
-      }
-    } else {
-      // When nullable and no value, keep input empty instead of defaulting to '00:00:00'
-      setLocalTimeValue(nullable ? '' : '00:00:00');
-    }
-  }, [value, nullable]);
-
-  const showClear = nullable && !disabled && value != null && value !== '';
-
-  const handleClear = (e: React.MouseEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    onTimeChange('');
-  };
-
-  const handleTimeChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const newTimeValue = e.target.value;
-      setLocalTimeValue(newTimeValue);
-    },
-    []
-  );
-
-  const handleTimeBlur = useCallback(() => {
-    // When input loses focus, update the parent
-    onTimeChange(localTimeValue);
-  }, [localTimeValue, onTimeChange]);
-
-  const handleKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      // When user presses Enter, update the parent
-      if (e.key === 'Enter') {
-        onTimeChange(localTimeValue);
-      }
-    },
-    [localTimeValue, onTimeChange]
-  );
-
-  return (
-    <div className="relative w-full select-none" data-testid={dataTestId}>
-      <div
-        className={cn(
-          'relative flex items-center rounded-md border border-input shadow-sm focus-within:ring-1 focus-within:ring-ring',
-          invalid && inputStyles.invalidInput
-        )}
-      >
-        <Clock
-          className={cn(
-            'ml-3 shrink-0',
-            dateTimeInputIconVariants({ scale }),
-            disabled && 'opacity-50'
-          )}
-        />
-        <Input
-          type="time"
-          step="1"
-          scale={scale}
-          value={localTimeValue}
-          onChange={handleTimeChange}
-          onBlur={handleTimeBlur}
-          onKeyDown={handleKeyDown}
-          disabled={disabled}
-          placeholder={placeholder || 'Select time'}
-          className={cn(
-            'bg-transparent appearance-none [&::-webkit-calendar-picker-indicator]:hidden cursor-pointer w-full border-0 shadow-none focus-visible:ring-0',
-            dateTimeInputTextVariants({ scale }),
-            invalid && inputStyles.invalidInput,
-            disabled && 'cursor-not-allowed',
-            showClear && invalid ? 'pr-16' : showClear || invalid ? 'pr-8' : ''
-          )}
-        />
-      </div>
-      {/* Icons absolutely positioned */}
-      {(showClear || invalid) && (
-        <div className="absolute right-2 top-1/2 -translate-y-1/2 flex flex-row items-center gap-1">
-          {showClear && (
-            <button
-              type="button"
-              tabIndex={-1}
-              aria-label="Clear"
-              onClick={handleClear}
-              className="p-1 rounded hover:bg-accent focus:outline-none cursor-pointer"
-            >
-              <X
-                className={cn(
-                  dateTimeInputIconVariants({ scale }),
-                  'text-muted-foreground hover:text-foreground'
-                )}
-              />
-            </button>
-          )}
-          {/* Invalid icon - rightmost */}
-          {invalid && (
-            <InvalidIcon message={invalid} className="pointer-events-auto" />
-          )}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const VariantComponents = {
-  Date: DateVariant,
-  DateTime: DateTimeVariant,
-  Time: TimeVariant,
-};
-
-export const DateTimeInputWidget: React.FC<DateTimeInputWidgetProps> = ({
-  id,
-  value,
-  placeholder,
-  disabled = false,
-  variant = 'Date',
-  nullable = false,
-  invalid,
-  format: formatProp,
-  scale = Scales.Medium,
-  'data-testid': dataTestId,
-}) => {
-  const eventHandler = useEventHandler();
-
-  // Normalize undefined to null when nullable
-  const normalizedValue = nullable && value === undefined ? undefined : value;
-
-  const handleDateChange = useCallback(
-    (selectedDate: Date | undefined) => {
-      if (disabled) return;
-      const isoString = selectedDate?.toISOString();
-      eventHandler('OnChange', id, [isoString]);
-    },
-    [disabled, eventHandler, id]
-  );
-
-  const handleTimeChange = useCallback(
-    (time: string) => {
-      if (disabled) return;
-
-      // For Time variant, send the time string directly
-      if (variant === 'Time') {
-        eventHandler('OnChange', id, [time]);
-      } else {
-        // For other variants, create a date with the selected time
-        const [hours, minutes, seconds] = time.split(':').map(Number);
-        const newDateTime = new Date();
-        newDateTime.setHours(hours, minutes, seconds);
-
-        eventHandler('OnChange', id, [newDateTime.toISOString()]);
-      }
-    },
-    [disabled, eventHandler, id, variant]
-  );
-
-  const VariantComponent = useMemo(() => VariantComponents[variant], [variant]);
-
-  return (
-    <VariantComponent
-      id={id}
-      value={normalizedValue}
-      placeholder={placeholder}
-      disabled={disabled}
-      nullable={nullable}
-      invalid={invalid}
-      format={formatProp}
-      scale={scale}
-      onDateChange={handleDateChange}
-      onTimeChange={handleTimeChange}
-      data-testid={dataTestId}
-    />
-  );
-};
